initial

+++ cifar100.py
@@ -93,7 +93,8 @@ def main():
     model = ResNet50Classifier()
     # trainer = Trainer(accelerator='npu', devices='0,1', max_epochs=5, strategy='deepspeed', precision=16)
     # trainer = Trainer(accelerator='npu', devices='0,1', max_epochs=5, strategy='deepspeed')
-    trainer = Trainer(accelerator='npu', devices='0,1', max_epochs=5, precision=16)
+    # trainer = Trainer(accelerator='npu', devices='0,1', max_epochs=5, precision=16)
+    trainer = Trainer(accelerator='npu', devices='0,1', max_epochs=5)
     trainer.fit(model, datamodule=data_module)
     trainer.test(model, datamodule=data_module)
